patricia: avoid left-shifting a negative value
mask is a uint8_t, so it promotes to int and (- mask) is negative whenever mask is non-zero, which is every time these lines are reached: the zero case is handled separately above each of them. Left-shifting a negative value is undefined per C99 6.5.7p4, and it is what UBSan reports as a left shift of a negative value, on a path the chunk cache uses constantly.

Negating as unsigned is defined as modular arithmetic and the left shift then simply discards the bits shifted out, producing the same bit pattern. Checked for all eight values mask can take; the result is identical in each.

+++ lib/datastruct/patricia.c
@@ -181,7 +181,7 @@ compare(const struct pnode * n, const uint8_t * key, size_t keylen,
 	}
 
 	/* Compute how the top bits differ. */
-	mm = (n->high ^ key[i]) & ((- n->mask) << 1);
+	mm = (n->high ^ key[i]) & ((-(unsigned int)n->mask) << 1);
 
 	/* If the top bits match, the node is a prefix of the key. */
 	if (mm == 0)
@@ -354,7 +354,7 @@ patricia_insert(PATRICIA * P, const uint8_t * key, size_t keylen, void * rec)
 				pnew->high = 0;
 				pnew->right = pnew2;
 			} else {
-				pnew->high = key[mlen] & ((- mask) << 1);
+				pnew->high = key[mlen] & ((-(unsigned int)mask) << 1);
 
 				/*
 				 * This looks wrong, but it actually works:
